OP-3682 - extracted AddonInfo to separate file
Parsing of v4 payload info will be required in Dependencies tool also.
